moved data modules to params.py
