+++ src/nemos/pp_glm/params.py
@@ -1,6 +1,8 @@
-from typing import Callable, Tuple, Union
+from typing import Callable
 
+import equinox as eqx
 import jax.numpy as jnp
+from jaxtyping import Array, Float, Int
 
 from ..glm.params import GLMParams
 from ..params import ModelParams
@@ -17,3 +19,23 @@ class PPGLMParamsWithKey(ModelParams):
         list[Callable[["PPGLMParamsWithKey"], jnp.ndarray | dict]]
     ):
         return [lambda p: p.params.coef]
+
+
+class X_ppglm(eqx.Module):
+    """Preprocessed predictors for PP-GLM."""
+
+    times: Float[Array, "n_events"]
+    ids: Int[Array, "n_events"]
+
+
+class y_ppglm(eqx.Module):
+    """Preprocessed spikes for PP-GLM."""
+
+    times: Float[Array, "n_spikes"]
+    ids: Int[Array, "n_spikes"]
+    idx: Int[Array, "n_spikes"]
+
+
+class mc_sample_ppglm(eqx.Module):
+    times: Float[Array, "n_samples"]
+    idx: Int[Array, "n_samples"]
